Редактирование + статус

+++ newsm.js
@@ -21,7 +21,6 @@ module.exports.newscl = class newscl{
         this.mpunixtime = results.mpdate
         this.date = results.date
         this.pictures = results.pictures
-        console.log('ИЗ БАЗЫ'+this.pictures);
         this.news_path_n = this.curpath(results.path)
     }
     
@@ -50,20 +49,23 @@ module.exports.newscl = class newscl{
         }
         keywords.forEach(element => {
         //console.log(element);
-            let indf = arr[i].indexOf(element)
-            if (indf!=-1) {
-            // console.log("OK" +arr[i] + " " + element); 
-                arr.splice(i, 1);
-            } else {
-            // console.log("NOOK" +arr[i] + " " + element); 
-            // arr.splice(i, i); // начиная с позиции 1, удалить 1 элемент
-            }
+                try {
+                    let indf = arr[i].indexOf(element)
+                    if (indf!=-1) {
+                    // console.log("OK" +arr[i] + " " + element); 
+                        arr.splice(i, 1);
+                    } else {
+                    // console.log("NOOK" +arr[i] + " " + element); 
+                    // arr.splice(i, i); // начиная с позиции 1, удалить 1 элемент
+                    }
+                } catch (error) {
+                    console.log('Ошибка indexOf 63 newsjs '+arr[i]);
+                }
             });
             
     }
     arr.unshift(dat);
     this.newsname= arr.join(' ');
-    console.log(this.newsname);
     return this.newsname
     //console.log(arr);
     }
@@ -76,7 +78,7 @@ module.exports.newscl = class newscl{
         let data_arr = mpdate.split(' ');
         let mpunixtime = parseInt((new Date('2022.'+data_arr[1]+'.'+data_arr[0]).getTime() / 1000).toFixed(0))
         if (isNaN(mpunixtime)) {mpunixtime = -1}
-        console.log(mpunixtime);
+        console.log('UNIX время: '+mpunixtime);
         return mpunixtime
     }
 
@@ -89,7 +91,7 @@ module.exports.newscl = class newscl{
     this.status = status
     this.mpunixtime = this.getmpunixtime(mpdate)
     this.pictures = pictures
-    console.log(news_path_n);
+    console.log('Путь к новости:'+news_path_n);
     this.news_path_n = news_path_n
     this.mpdate = mpdate
     this.save();
@@ -98,26 +100,41 @@ module.exports.newscl = class newscl{
     edit(){
         //Удалить фото
         //console.log(this.pictures);
+
+
         const dbworker = new this.db.dbworker(); 
         this.pictures.forEach(element => {
             dbworker.addpict(this.id,element)
             });
         this.deldocx(this.news_path_n)
+        this.mpunixtime = this.getmpunixtime(this.mpdate)
+        dbworker.updatenews(this.id,this.head,this.cont,this.autor,this.mpunixtime,this.news_path_n)
 
-        dbworker.updatenews(this.id,this.head,this.cont,this.autor,this.status,this.mpunixtime,this.news_path_n)
+    }
+
+    renamefolder(){
+        const dbworker = new this.db.dbworker(); 
+        console.log('Старая папка из бд: ' + JSON.parse(this.news_path_n));
+        this.fs.renameSync(JSON.parse(this.news_path_n),this.pathmod.join(this.news_path,'new',this.newsname))
+        console.log('Новое имя: '+this.pathmod.join(this.news_path,'new',this.newsname));
+        this.news_path_n = JSON.stringify(this.pathmod.join(this.news_path,'new',this.newsname));
+        dbworker.updatefolder(this.id,this.news_path_n)
     }
 
     delpicture(picfordel){
         const dbworker = new this.db.dbworker(); 
         picfordel=picfordel.split(',');
-        console.log(picfordel);
+        console.log('Имена фоток для удаления из формы:'+picfordel);
         picfordel.forEach(element => {
             element = element.split('/');
             element = element[element.length-1]
-            let does = this.fs.removeSync(this.pathmod.join(this.news_path_n,element))
-            console.log(dbworker.dellpict(this.id,element));
-            console.log(does+' Фалы удалены: '+element);
-            console.log("Имя на удаление " +element);
+            console.log(element);
+            if (element!='') {
+                console.log('Удаляю: '+this.pathmod.join(this.news_path_n,element));
+                let does = this.fs.removeSync(this.pathmod.join(this.news_path_n,element))
+                dbworker.dellpict(this.id,element);
+                console.log('Файл удален: '+element);
+            }
         });
         
 
@@ -130,24 +147,30 @@ module.exports.newscl = class newscl{
     try {
         this.news_path_n = JSON.stringify(this.news_path_n)
         this.getname(this.mpdate,this.head)
-        console.log(par);
+        //this.renamefolder()
+        console.log('Сохрнаить?'+par);
         if (par!=undefined) {
+            this.renamefolder()
+           // this.news_path_n = JSON.stringify(this.news_path_n)
+            this.getname(this.mpdate,this.head)
             this.edit()
         } else {
             dbworker.addnews(this.head,this.cont,this.autor,this.status,this.mpunixtime,this.unixTime(new Date()),this.pictures,this.news_path_n);  
             }
             doxworker.addnews(this.newsname, this.head, this.cont ,this.news_path_n); //Заменить первое на имя документа
+            console.log('Название:'+this.newsname);
     } catch (err) {
-        console.log('DATABASE '+err);
+        console.log('Ошибка 144 строка: '+err);
         return false;
         }
+        
         return true;
     }
     deldocx(dirpath) {
-        console.log('deleteron');
+
         dirpath = this.curpath(dirpath)
         let files = this.fs.readdirSync(dirpath)
-        console.log(files);
+        console.log('Файлы в диерктории удаления:'+files);
         files.forEach(files => {
             let arr = files.split('.');
             if (arr[arr.length-1]=='docx'){
@@ -155,6 +178,13 @@ module.exports.newscl = class newscl{
                 console.log(does+' Фалы удалены: '+files);
             }
         });
+    }
+    getdbfolder(){
+        let foldern = this.news_path_n.split('\\');
+       // foldern = foldern[foldern.length-1];
+        console.log(foldern);
+        foldern = this.pathmod.join(this.news_path,'new',foldern)
+        return foldern;
     }
 }
 const { dbworker } = require('./db.js');
